Optimisation step for getting id, data and extra_data.
I removed the Entity and EntityBed structures, that were used to store information about entities. And immediately store the colour as an uint16_t, so it should be faster to retrieve this during
rendering.

I also added Chunk::getBlockInfo, which expects a position, and the addresses of where to store the id, data and extra_data values. This is an optimisation, as Chunk::getBlockExtraData needs the
block id, as the default extra_data can be different between blocks. Chunk::getBlockInfo gets the id, as before, stores it in the caller of Chunk::getBlockInfo, and passes that value to
Chunk::getBlockExtraData. Thus Chunk::getBlockExtraData does not need to calculate the block id again.